requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.0
dune-client==1.3.0
psycopg2-binary==2.9.6
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
pytest==7.4.0
pytest-mock==3.11.1
pendulum==2.1.2
dbt-core==1.8.4
dbt-postgres==1.8.2
pytest-dbt-core==0.2.4
SQLAlchemy==1.4.49
python-dateutil==2.8.2
flake8==6.0.0
black==23.3.0
flask-session==0.5.0
python-dotenv>=0.20.0
pre-commit==2.17.0
pytest-cov==3.0.0
pylint==2.17.4